Fix reset code.
The HW manual states that performing a module reset whilst the VEU
is still operating can cause the HW to stop working. The correct
way to reset the VEU is to perform a software reset if the VEU is
running, wait for this to complete and clear the VEU end interrupt.