vio_blend: Enable setting of the virtual BRU input as master layer
When compositing several layers into one with the BRU, a master layer
must be specified and the offset of all other layers are indicated
in relation to the top left corner of the master layer an should not
extend past the bottom right corner of the master layer.  If none of
the layers to be composited can bound all of the other layers, the
virtual layer can be used as the master layer.

The virtual layer, when enabled, is filled with opaque black and
is at the bottom of the blend stack.