vio_blend; Support surface resizing while blending
Resize a surface to the size specified in .blend_out before
blending.

Note: Perfoming surface resizes before blending depends on the
availablility of a scaling module.  The number of scaling modules
available my be less than the number of possible blending surfaces,
so it may not be possible to resize all surfaces when blending.